Refactor part 1
Refactor everything except send/recv/stat (for now)
